+++ scripts/test.sh
@@ -12,5 +12,4 @@ cargo test --manifest-path=opentelemetry/Cargo.toml --all-features -- --ignored 
 
 cargo test --manifest-path=opentelemetry/Cargo.toml --all-features
 cargo test --manifest-path=opentelemetry-jaeger/Cargo.toml --all-features -- --test-threads=1
-cargo test --manifest-path=opentelemetry-otlp/Cargo.toml --features "trace,grpc-sys" --no-default-features
 cargo test --manifest-path=opentelemetry-zipkin/Cargo.toml --all-features